# doc/openapi.yaml
openapi: 3.0.3
info:
  title: JuYu AI Platform API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
...
  /execute:
    post:
      summary: Execute a scene task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Request'
      responses:
        '200':
          description: Task accepted or completed
  /tasks:
    get:
      summary: List tasks
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: scene
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer }
        - in: query
          name: offset
          schema: { type: integer }
      responses:
        '200':
          description: Task list
...
components:
  schemas:
    Request:
      type: object
      properties:
        scene:
          type: string
        input:
          type: string
        payload:
          type: object
          additionalProperties: true